rag-wiki-python/requirements.txt
# 智维Wiki Python AI服务依赖
# 核心框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM & RAG
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
sentence-transformers>=2.2.0
transformers>=4.36.0
torch>=2.0.0

# 向量数据库
pymilvus>=2.4.0
qdrant-client>=1.7.0

# 异步任务
celery[redis]>=5.3.0
redis>=5.0.0

# 存储
minio>=7.2.0

# 数据库
pymysql>=1.1.0
sqlalchemy>=2.0.0

# HTTP客户端
httpx>=0.25.0
aiohttp>=3.9.0

# 重试与容错
tenacity>=8.2.0

# 链路追踪
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-httpx>=0.43b0

# 文件处理
python-multipart>=0.0.6

# 认证
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# 文档解析
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.22
pandas>=2.1.0
numpy>=1.26.0

# OCR (可选，体积较大)
# paddleocr>=2.7.0
# paddlepaddle>=2.5.0

# 代码沙箱
docker>=6.1.0

# 监控
sentry-sdk[fastapi]>=1.39.0
prometheus-client>=0.20.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.12.0
isort>=5.13.0
